# slots-game/server/openapi.yaml
openapi: 3.1.0
...
info:
  title: Slots Game RGS Integration API
  version: 1.0.0
  summary: Operator and player-client contract for the clean-room RGS.
  description: |
    This draft documents the production RGS HTTP boundary implemented under
    `server/internal/rgsapi`. All money, revision and sequence values are
    canonical base-10 strings so JavaScript clients cannot lose integer
    precision. Unknown and duplicate JSON fields are rejected.

    Operator routes use the repository's fixed Ed25519 HTTP Message Signature
    profiles for requests and responses. Client routes use a short-lived
    RGS-ACCESS v2 compact bearer token issued after a one-time launch-code
    exchange and replaced only while the prior token and durable session remain
    valid. Signed endpoints do not accept query strings.

    This document is not evidence of regulatory approval or certification. A
    real-money deployment additionally requires the applicable licences,
    independent RNG and mathematics certification, an independent security
    audit, and a wallet adapter plus conformance approval for every operator.
servers:
  - url: https://rgs.example.invalid
    description: Replace with the operator-approved RGS origin.
...
paths:
  /operator/v1/launches:
    post:
      tags: [Operator]
      operationId: createLaunch
      summary: Create one short-lived, single-use launch code
      description: |
        The authenticated operator comes from the verified signing key. The
        request binds a player and wallet session to one immutable game
        definition, currency and jurisdiction. Idempotency-Key identifies one
        browser handoff, independently of the durable `sessionId`. Exact retries
        use the same key and replay the same launch code and original expiry,
        including after consumption or exchange expiry. That response is never
        reactivated. Its digest remains an idempotency tombstone until 25 hours
        after `expiresAt`. A new key issues a new code for the same active,
        unexpired session without resetting its balance, absolute expiry,
        revision, or feature state. For an existing session `balanceMinor` and
        `sessionTtlSeconds` are creation-only inputs and do not replace durable
        state. Within the same operator/session scope, reusing a key with
        different request data is rejected. Every
        response after the pre-authentication capacity gate is signed when a
        tenant response key can be selected. At process capacity, the gate may
        return a generic unsigned, data-free 503 CAPACITY_UNAVAILABLE with
        Retry-After before identity, signature, or nonce parsing. It is transport
        backpressure, not an authenticated business result; retain the body and
        idempotency key and follow the uncertain-transport retry/status contract
        with a fresh nonce and signature. If no safe tenant key can be selected,
        the server instead fails closed with an unsigned, data-free 503
        RESPONSE_SIGNING_UNAVAILABLE envelope.
      security:
        - operatorMessageSignature: []
      parameters:
        - $ref: '#/components/parameters/OperatorId'
        - $ref: '#/components/parameters/SignedRequestId'
        - $ref: '#/components/parameters/Nonce'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ContentDigest'
        - $ref: '#/components/parameters/SignatureInput'
        - $ref: '#/components/parameters/Signature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLaunchRequest'
      responses:
        '201':
          description: Launch credential created, or the exact retained response replayed for the same handoff key.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Content-Digest:
              $ref: '#/components/headers/ContentDigest'
            Signature-Input:
              $ref: '#/components/headers/SignatureInput'
            Signature:
              $ref: '#/components/headers/Signature'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchSuccessEnvelope'
        '400':
          $ref: '#/components/responses/SignedError'
        '401':
          $ref: '#/components/responses/SignedError'
        '403':
          $ref: '#/components/responses/SignedError'
        '409':
          $ref: '#/components/responses/SignedError'
        '413':
          $ref: '#/components/responses/SignedError'
        '415':
          $ref: '#/components/responses/SignedError'
        '405':
          $ref: '#/components/responses/SignedMethodNotAllowed'
        '410':
          $ref: '#/components/responses/SignedError'
        '423':
          $ref: '#/components/responses/SignedError'
        '429':
          $ref: '#/components/responses/SignedRateLimited'
        '500':
          $ref: '#/components/responses/SignedError'
        '503':
          $ref: '#/components/responses/OperatorUnavailable'
...
components:
  securitySchemes:
    operatorMessageSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        Fixed Ed25519 request profile. All X-Operator-Id, X-Request-Id,
        X-Nonce, Idempotency-Key, Content-Digest, Signature-Input and Signature
        headers are mandatory. It covers "@method", "@authority", "@path",
        "content-digest", "content-type", "x-operator-id", "x-request-id",
        "x-nonce" and "idempotency-key" in that order.
    clientAccessToken:
      type: http
      scheme: bearer
      bearerFormat: RGS-ACCESS-v2
      description: Fixed compact Ed25519 v2 token, including game-definition hash; not a general JWT profile.
    operationsBearer:
      type: http
      scheme: bearer
      bearerFormat: RGS-OPERATIONS
      description: A deployment-only secret loaded from RGS_OPERATIONS_BEARER_TOKEN_FILE; required for readiness and metrics in production and on every non-loopback operations listener.
  parameters:
    OperatorId:
      name: X-Operator-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
      description: Must match the tenant bound to the verified key.
    SignedRequestId:
      name: X-Request-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
      description: Signed correlation ID; must not contain a player identifier.
...
    Nonce:
      name: X-Nonce
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 256
      description: Cryptographically random, single-use nonce within the verified key scope.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
      description: Stable across retries of the same operator operation. For createLaunch it identifies one browser handoff, not the durable session; its response is retained until 25 hours after code expiry. Use a new value only when intentionally issuing the next handoff.
    ContentDigest:
      name: Content-Digest
      in: header
      required: true
      schema:
        type: string
        pattern: '^sha-256=:[A-Za-z0-9+/]+={0,2}:$'
    SignatureInput:
      name: Signature-Input
      in: header
      required: true
      schema:
        type: string
      description: Exact sig1 components, created/expires, keyid and alg="ed25519".
    Signature:
      name: Signature
      in: header
      required: true
      schema:
        type: string
        pattern: '^sig1=:[A-Za-z0-9+/]+={0,2}:$'
  headers:
    RequestId:
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
    ContentDigest:
      required: true
      schema:
        type: string
        pattern: '^sha-256=:[A-Za-z0-9+/]+={0,2}:$'
    SignatureInput:
      required: true
      schema:
        type: string
    Signature:
      required: true
      schema:
        type: string
        pattern: '^sig1=:[A-Za-z0-9+/]+={0,2}:$'
    RetryAfter:
      schema:
        type: integer
        minimum: 1
        maximum: 60
  responses:
    SignedError:
      description: Signed error envelope. Inspect the stable body code; no internal error text is exposed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Content-Digest:
          $ref: '#/components/headers/ContentDigest'
        Signature-Input:
          $ref: '#/components/headers/SignatureInput'
        Signature:
          $ref: '#/components/headers/Signature'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    SignedMethodNotAllowed:
      description: Signed method-not-allowed error envelope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Content-Digest:
          $ref: '#/components/headers/ContentDigest'
        Signature-Input:
          $ref: '#/components/headers/SignatureInput'
        Signature:
          $ref: '#/components/headers/Signature'
        Allow:
          required: true
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    SignedRateLimited:
      description: Signed local admission-control rejection.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Content-Digest:
          $ref: '#/components/headers/ContentDigest'
        Signature-Input:
          $ref: '#/components/headers/SignatureInput'
        Signature:
          $ref: '#/components/headers/Signature'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    OperatorUnavailable:
      description: |
        Dependency, capacity, or response-signing failure. Normally signed. The
        process-wide pre-authentication capacity gate returns an intentionally
        unsigned CAPACITY_UNAVAILABLE response with Retry-After because no tenant
        identity is trusted yet. If a safe tenant response key cannot be selected,
        the minimal RESPONSE_SIGNING_UNAVAILABLE envelope is also intentionally
        unsigned. Neither unsigned response is an authoritative business outcome:
        after Retry-After, query status where available or retain the exact body
        and idempotency key and retry with a fresh nonce/request ID/signature. Do
        not infer nonce or round side effects and do not create a new business
        identity to bypass capacity.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Content-Digest:
          schema:
            type: string
            pattern: '^sha-256=:[A-Za-z0-9+/]+={0,2}:$'
          description: Absent for the pre-authentication capacity gate and when response signing itself was unavailable.
        Signature-Input:
          schema:
            type: string
          description: Absent for the pre-authentication capacity gate and when response signing itself was unavailable.
        Signature:
          schema:
            type: string
            pattern: '^sig1=:[A-Za-z0-9+/]+={0,2}:$'
          description: Absent for the pre-authentication capacity gate and when response signing itself was unavailable.
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Identifier:
      type: string
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
      examples: [req_01JABCDEF]
    Digest:
      type: string
      pattern: '^[a-f0-9]{64}$'
      examples: [0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef]
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      examples: [EUR]
    Jurisdiction:
      type: string
      pattern: '^[A-Z0-9][A-Z0-9-]{1,15}$'
      examples: [GB]
    MoneyMinor:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
      maxLength: 19
      description: Canonical non-negative signed-int64 value in minor currency units.
      examples: ['100']
...
    CreateLaunchRequest:
      type: object
      additionalProperties: false
      required:
        - playerId
        - walletAccountId
        - walletSessionId
        - sessionId
        - gameId
        - definitionVersion
        - definitionHash
        - currency
        - currencyExponent
        - jurisdiction
        - balanceMinor
        - sessionTtlSeconds
      properties:
        playerId:
          $ref: '#/components/schemas/Identifier'
        walletAccountId:
          $ref: '#/components/schemas/Identifier'
        walletSessionId:
          $ref: '#/components/schemas/Identifier'
        sessionId:
          $ref: '#/components/schemas/Identifier'
        gameId:
          $ref: '#/components/schemas/Identifier'
        definitionVersion:
          $ref: '#/components/schemas/Identifier'
        definitionHash:
          $ref: '#/components/schemas/Digest'
        currency:
          $ref: '#/components/schemas/Currency'
        currencyExponent:
          type: integer
          minimum: 0
          maximum: 6
        jurisdiction:
          $ref: '#/components/schemas/Jurisdiction'
        balanceMinor:
          $ref: '#/components/schemas/MoneyMinor'
          description: Creation-only balance snapshot. Relaunch keeps the persisted session balance and ignores this value for durable state; retries with one Idempotency-Key must still keep the complete request unchanged.
        sessionTtlSeconds:
          type: integer
          minimum: 60
          maximum: 86400
          description: Creation-only lifetime. Relaunch keeps the persisted absolute session expiry and never extends it; retries with one Idempotency-Key must still keep the complete request unchanged.
...
    LaunchSuccessEnvelope:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/LaunchResponse'
        requestId:
          $ref: '#/components/schemas/Identifier'
    LaunchResponse:
      type: object
      additionalProperties: false
      required: [launchCode, exchangeUrl, expiresAt]
      properties:
        launchCode:
          type: string
          pattern: '^lc_[A-Za-z0-9_-]{43}$'
          minLength: 46
          maxLength: 46
        exchangeUrl:
          type: string
          format: uri
          pattern: '^https://'
        expiresAt:
          type: string
          format: date-time
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          $ref: '#/components/schemas/Identifier'
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,127}$'
        message:
          type: string
          minLength: 1
          maxLength: 256
          description: Stable non-sensitive guidance; never an internal error string.
...